# Serialization performance improvements