--- a/src/System.Text.Json/src/System/Text/Json/Serialization/WriteStackFrame.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/WriteStackFrame.cs
@@ -3,8 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections;
-using System.Collections.Generic;
 using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using System.Text.Json.Serialization;
 
 namespace System.Text.Json
@@ -32,8 +32,8 @@ namespace System.Text.Json
 
         // The current property.
         public bool PropertyEnumeratorActive;
+        public int PropertyEnumeratorIndex;
         public ExtensionDataWriteStatus ExtensionDataStatus;
-        public Dictionary<string, JsonPropertyInfo>.Enumerator PropertyEnumerator;
         public JsonPropertyInfo JsonPropertyInfo;
 
         public void Initialize(Type type, JsonSerializerOptions options)
@@ -111,7 +111,7 @@ namespace System.Text.Json
             ExtensionDataStatus = ExtensionDataWriteStatus.NotStarted;
             IsIDictionaryConstructible = false;
             JsonClassInfo = null;
-            PropertyEnumerator = default;
+            PropertyEnumeratorIndex = 0;
             PropertyEnumeratorActive = false;
             PopStackOnEndCollection = false;
             PopStackOnEndObject = false;
@@ -119,6 +119,7 @@ namespace System.Text.Json
             EndProperty();
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void EndProperty()
         {
             IsIDictionaryConstructibleProperty = false;
@@ -139,20 +140,28 @@ namespace System.Text.Json
             PopStackOnEndCollection = false;
         }
 
+        // AggressiveInlining used although a large method it is only called from one location and is on a hot path.
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void NextProperty()
         {
             EndProperty();
 
             if (PropertyEnumeratorActive)
             {
-                if (PropertyEnumerator.MoveNext())
+                int len = JsonClassInfo.PropertyCacheArray.Length;
+                if (PropertyEnumeratorIndex < len)
                 {
+                    if ((PropertyEnumeratorIndex == len - 1) && JsonClassInfo.DataExtensionProperty != null)
+                    {
+                        ExtensionDataStatus = ExtensionDataWriteStatus.Writing;
+                    }
+
+                    PropertyEnumeratorIndex++;
                     PropertyEnumeratorActive = true;
                 }
                 else
                 {
                     PropertyEnumeratorActive = false;
-                    ExtensionDataStatus = ExtensionDataWriteStatus.Writing;
                 }
             }
             else